packages/cli: rename helpers to lib
