+++ packages/cli/src/commands/app/build.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { run } from 'helpers/run';
+import { run } from 'lib/run';
 
 export default async () => {
   const args = ['build'];
