+++ packages/cli/src/commands/build-cache/cache.ts
@@ -16,11 +16,11 @@
 
 import fs from 'fs-extra';
 import { resolve as resolvePath, relative as relativePath } from 'path';
-import { runPlain, runCheck } from 'helpers/run';
+import { runPlain, runCheck } from 'lib/run';
 import { Options } from './options';
 import { extractArchive, createArchive } from './archive';
-import { paths } from 'helpers/paths';
-import { version, isDev } from 'helpers/version';
+import { paths } from 'lib/paths';
+import { version, isDev } from 'lib/version';
 
 const INFO_FILE = '.backstage-build-cache';
