+++ packages/cli/src/commands/clean/clean.ts
@@ -17,7 +17,7 @@
 import fs from 'fs-extra';
 import { resolve as resolvePath, relative as relativePath } from 'path';
 import { getDefaultCacheOptions } from 'commands/build-cache/options';
-import { paths } from 'helpers/paths';
+import { paths } from 'lib/paths';
 
 export default async function clean() {
   const cacheOptions = getDefaultCacheOptions();
