+++ packages/cli/src/commands/create-plugin/createPlugin.ts
@@ -26,9 +26,9 @@ import {
   addCodeownersEntry,
   getCodeownersFilePath,
 } from './lib/codeowners';
-import { paths } from 'helpers/paths';
-import { version } from 'helpers/version';
-import { Task, templatingTask } from 'helpers/tasks';
+import { paths } from 'lib/paths';
+import { version } from 'lib/version';
+import { Task, templatingTask } from 'lib/tasks';
 const exec = promisify(execCb);
 
 async function checkExists(rootDir: string, id: string) {
