+++ packages/cli/src/commands/remove-plugin/removePlugin.test.ts
@@ -17,7 +17,7 @@
 import fse from 'fs-extra';
 import path from 'path';
 import os from 'os';
-import { paths } from '../../helpers/paths';
+import { paths } from 'lib/paths';
 import {
   addExportStatement,
   capitalize,
