+++ packages/cli/src/commands/remove-plugin/removePlugin.ts
@@ -18,8 +18,8 @@ import path from 'path';
 import chalk from 'chalk';
 import inquirer, { Answers, Question } from 'inquirer';
 import { getCodeownersFilePath } from '../create-plugin/lib/codeowners';
-import { paths } from 'helpers/paths';
-import { Task } from 'helpers/tasks';
+import { paths } from 'lib/paths';
+import { Task } from 'lib/tasks';
 // import os from 'os';
 
 const BACKSTAGE = '@backstage';
