+++ packages/cli/src/index.ts
@@ -16,8 +16,8 @@
 
 import program from 'commander';
 import chalk from 'chalk';
-import { exitWithError } from 'helpers/errors';
-import { version } from 'helpers/version';
+import { exitWithError } from 'lib/errors';
+import { version } from 'lib/version';
 
 const main = (argv: string[]) => {
   program.name('backstage-cli').version(version);
