+++ packages/cli/src/lib/errors.ts
