+++ packages/cli/src/lib/paths.ts
