+++ packages/cli/src/lib/run.ts
