+++ packages/cli/src/lib/version.ts
